1. Worked_by_Seungwon 브랜치 생성   2. 앞으로 필요한 추가 이미지 파일 추가, 폴더로 정리   3. 외벽과 충돌체크   4. 프로젝트이름 변경
