--- 10-2 슈팅게임.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30611.23
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "10-2 슈팅게임", "10-2 슈팅게임.vcxproj", "{A6074820-AFE3-4655-AF9F-A4C7443EB292}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A6074820-AFE3-4655-AF9F-A4C7443EB292}.Debug|x64.ActiveCfg = Debug|x64
-		{A6074820-AFE3-4655-AF9F-A4C7443EB292}.Debug|x64.Build.0 = Debug|x64
-		{A6074820-AFE3-4655-AF9F-A4C7443EB292}.Debug|x86.ActiveCfg = Debug|Win32
-		{A6074820-AFE3-4655-AF9F-A4C7443EB292}.Debug|x86.Build.0 = Debug|Win32
-		{A6074820-AFE3-4655-AF9F-A4C7443EB292}.Release|x64.ActiveCfg = Release|x64
-		{A6074820-AFE3-4655-AF9F-A4C7443EB292}.Release|x64.Build.0 = Release|x64
-		{A6074820-AFE3-4655-AF9F-A4C7443EB292}.Release|x86.ActiveCfg = Release|Win32
-		{A6074820-AFE3-4655-AF9F-A4C7443EB292}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FC23BE21-B5D8-4B92-A026-51EC08F822AD}
-	EndGlobalSection
-EndGlobal
