+++ Bomberman_project.rc
@@ -1,0 +1,89 @@
+// Microsoft Visual C++ generated resource script.
+//
+#include "resource.h"
+
+#define APSTUDIO_READONLY_SYMBOLS
+/////////////////////////////////////////////////////////////////////////////
+//
+// Generated from the TEXTINCLUDE 2 resource.
+//
+#include "winres.h"
+
+/////////////////////////////////////////////////////////////////////////////
+#undef APSTUDIO_READONLY_SYMBOLS
+
+/////////////////////////////////////////////////////////////////////////////
+// 한국어(대한민국) resources
+
+#if !defined(AFX_RESOURCE_DLL) || defined(AFX_TARG_KOR)
+LANGUAGE LANG_KOREAN, SUBLANG_DEFAULT
+#pragma code_page(949)
+
+#ifdef APSTUDIO_INVOKED
+/////////////////////////////////////////////////////////////////////////////
+//
+// TEXTINCLUDE
+//
+
+1 TEXTINCLUDE 
+BEGIN
+    "resource.h\0"
+END
+
+2 TEXTINCLUDE 
+BEGIN
+    "#include ""winres.h""\r\n"
+    "\0"
+END
+
+3 TEXTINCLUDE 
+BEGIN
+    "\r\n"
+    "\0"
+END
+
+#endif    // APSTUDIO_INVOKED
+
+
+/////////////////////////////////////////////////////////////////////////////
+//
+// Bitmap
+//
+
+IDB_BITMAP1             BITMAP                  "maps_bmp/background.bmp"
+
+IDB_BITMAP2             BITMAP                  "objects_bmp/block.bmp"
+
+IDB_BITMAP3             BITMAP                  "characters_bmp/Isaac.bmp"
+
+IDB_BITMAP4             BITMAP                  "objects_bmp/tear.bmp"
+
+IDB_BITMAP5             BITMAP                  "monsters_bmp/monster.bmp"
+
+IDB_BITMAP6             BITMAP                  "characters_bmp/Magdalene.bmp"
+
+IDB_BITMAP7             BITMAP                  "characters_bmp/Lazarus.bmp"
+
+IDB_BITMAP8             BITMAP                  "characters_bmp/Samson.bmp"
+
+IDB_BITMAP9             BITMAP                  "characters_bmp\\Eve.bmp"
+
+IDB_BITMAP10            BITMAP                  "objects_bmp\\bomb.bmp"
+
+IDB_BITMAP11            BITMAP                  "objects_bmp\\item.bmp"
+
+#endif    // 한국어(대한민국) resources
+/////////////////////////////////////////////////////////////////////////////
+
+
+
+#ifndef APSTUDIO_INVOKED
+/////////////////////////////////////////////////////////////////////////////
+//
+// Generated from the TEXTINCLUDE 3 resource.
+//
+
+
+/////////////////////////////////////////////////////////////////////////////
+#endif    // not APSTUDIO_INVOKED
+
