+++ Bomberman_project.vcxproj.filters
@@ -1,0 +1,70 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <Filter Include="소스 파일">
+      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
+      <Extensions>cpp;c;cc;cxx;c++;cppm;ixx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
+    </Filter>
+    <Filter Include="헤더 파일">
+      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
+      <Extensions>h;hh;hpp;hxx;h++;hm;inl;inc;ipp;xsd</Extensions>
+    </Filter>
+    <Filter Include="리소스 파일">
+      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
+      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
+    </Filter>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="resource.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+  </ItemGroup>
+  <ItemGroup>
+    <ResourceCompile Include="Bomberman_project.rc">
+      <Filter>리소스 파일</Filter>
+    </ResourceCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <Image Include="map_bmp\background.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="objects_bmp\block.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="characters_bmp\Isaac.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="objects_bmp\tear.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="monsters_bmp\monster.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="characters_bmp\Magdalene.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="characters_bmp\Lazarus.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="characters_bmp\Samson.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="maps_bmp\background.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="characters_bmp\Eve.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="objects_bmp\bomb.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+    <Image Include="objects_bmp\item.bmp">
+      <Filter>리소스 파일</Filter>
+    </Image>
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="Bomberman_project.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+  </ItemGroup>
+</Project>
