--- Player.h
@@ -1,11 +1,0 @@
-#pragma once
-class Player
-{
-public:
-	int _id;
-	int _x, _y;
-	int _type;
-	int _heart;
-	int _bomb_count;
-};
-
