+++ resource.h
@@ -7,12 +7,18 @@
 #define IDB_BITMAP3                     105
 #define IDB_BITMAP4                     106
 #define IDB_BITMAP5                     108
+#define IDB_BITMAP6                     111
+#define IDB_BITMAP7                     112
+#define IDB_BITMAP8                     113
+#define IDB_BITMAP9                     115
+#define IDB_BITMAP10                    119
+#define IDB_BITMAP11                    120
 
 // Next default values for new objects
 // 
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE        109
+#define _APS_NEXT_RESOURCE_VALUE        121
 #define _APS_NEXT_COMMAND_VALUE         40001
 #define _APS_NEXT_CONTROL_VALUE         1001
 #define _APS_NEXT_SYMED_VALUE           101
